Fixed the collecting messages bug
Bug: raw contact id is hard to find in some cases, no id when sending a message. So
needed to do some extra queries to find the raw id which made process slower. (to 4secs)

A solution to fix this, but will not be implemented would be to add a second thread to seperate the case of collecting inbox and collectin sent messages.

At this moment contactInfo object contains phone numbers, messages, total message sent, and total message received. and timestamps for those messages.

TODO:
1. Need to implement the graph stage of the app now.
2. Comparsion tab also
3. add menu options, including changing contacts being graph and time period
4. add search and add contact feature.

NICE TO HAVE:
1. Add average # of words sent per message.
2. Popular words used from a contact
3. Average period between sent/received messages

These nice to haves will make both the graphs and comparison tab more useful and should not be too diffcult to implement if I am aware of them, and plan for them.

4. Add second thread! This is very important if I am going to add the nice to add features.