Collected yValues for contacts
Just need to get datasets for the ocntacts and graph them now
That should be it for the graph. Then should implement the menu options which will had a bit more functions like searching contacts then adding them, or clearing up, doesnt seem to hard really
Then need to move on to the card fragment.
This I don't think I am going to put much effort into maybe show some calculation that a customer may be interested in ratio of sent/recevied,
most popular word, timespan between message. This could be made well with alot of feature if time allowed