Fixed some bugs to do with graphing.
Need to rewrite the ContactLinkedList add function. It is
a piece of crap to understand and it does not work properly.
This is the next priority task to complete.

Also graph colours are currently at 10~ and need to be checked if all
colours are ok. Most likely will need to find better colours through
w3schools, should be simple enough.