Modified Graph fragment to be better. fixed some bugs.
1. Removed duplicate sames from legend
2. Added marker to graph, so that when A point is clicked, then a label pops up and displays the
contacts name, and value at that point that was clicked.
3. Fixed clipping issue with legend when there is too many contacts being graphed.
4. Fixed bug where different colours were being used for the same contact (Send messages dataset was 1
colour and received dataset was another ; it should just be 1 colour)

SIMS to Complete :

1. Do navigation menu to allow user to change graph values (interval, period, type of messgaes, top x)
Maybe also include a disable/enable legend
2. Implement the add contact to graph
3. Clear button (This may also just be in the navigation menu instead.)
4. Delete button. This will most likely need to be in the menu next to add. where you can search for contact
in the graph and then click a name to have it removed.
- One issue with this will be when both the receieved and sent messages are graphed. Or this might not be an issue

The Graph fragment will be considered finished when the above 4 SIMS are complete

Next will be the Card Fragment. I will post the tasks/SIMS that need to be done in a commit after the graph
fragment SIMS are complete.