Changed graph requests, fixed bugs, drop down menu change
Changed graph requests so that it takes in a list of contacts instead of
before where it took in a integer value for number of contacts to graph and
then made a request to get the list of contacts. This makes the contacts to graph
responsibility on the caller. (More flexiable with action bar buttons)

Fixed some bugs and gonna log in a bug if I can. First bug was with
choosing the week action in period, it was not formating properly.
Another bug was too many threads were being generated to graph the data
points and so I changed it to just use the main thread for now.
The data point collection/graphing needs to be a background thread, will file
and issue for this. Also I removed the minutes option because it is not
very useful for the user I believe and it also creates way to many datapoints for
a smartphone to handle efficiently.

Lastly, changed the action bar button drop down menu colour themes. This is to begin the
process to loading contacts to add/delete from the graph.

TODO :

1. Finish add/delete/clear action bar buttons
2. Complete Card Fragment (This may or may not be a hard process)

Nice to haves :
1. background thread(8 -> 8 cpus) to collect and graph data points (Don't know how hard this task is, need to research)
2. Add a donation option under Others. Not very important since it does not effect customers initially, but
will only effect whether or not I continue to support the app.
3. Add a turnOn/turnOff ads. Again same as 2. So not critcal for release. If This app generates enough user then
I will be inclined to refactor and clean up the UI and implements addition features