Implemented the Clear button and fixed bug
Forgot to include the clear button functionality. So implmented it.
I found a bug and fixed it also. The bug was when the ContactLinkedList
is empty and the first item added to the ContactLinkedList was then]
not added to the mapping and so the Contact could not be deleted if
added and then the user tried to deletge the contact.