Fixed the loading data... message bug, and prepared CardFragment
The CardFragment now is just going to be a list of the contacts being
graphed currently and some quick information about those contacts.

All the action bar buttons and nav menu change that I origanally had
planned is being scraped for a simpler and easy implementation of what
I had planned.

The code is still in place to make the switch to doing what I origanlly
wanted wit the CardFragment, but for now I think what I am going for
is 'Good Enough'